fix a warning: format '%d' expects type 'int', but argument 4 has type 'png_uint_32'
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@1413 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ src/texk/web2c/luatexdir/image/writepng.c
@@ -485,7 +485,7 @@ void copy_png(image_dict * idict)
                "/BitsPerComponent %i"
                "/Predictor 10>>\n>>\nstream\n", streamlength,
                info_p->color_type == 2 ? 3 : 1,
-               info_p->width, info_p->bit_depth);
+               (int)info_p->width, info_p->bit_depth);
     /* 2nd pass to copy data */
     endflag = false;
     if (fseek(fp, 8, SEEK_SET) != 0)
